requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
pymongo==4.5.0